refactored bundle automation to shared library
