+++ bundle-generation/bundles-to-charts.py
@@ -7,11 +7,12 @@ import argparse
 import os
 import shutil
 import yaml
-import array
 import logging
 import coloredlogs
 import sys
-import re
+
+import common
+
 from git import Repo, exc
 from packaging import version
 
@@ -19,66 +20,6 @@ from validate_csv import *
 
 # Configure logging with coloredlogs
 coloredlogs.install(level='DEBUG')  # Set the logging level as needed
-
-# Split a string at a specified delimiter.  If delimiter doesn't exist, consider the
-# string to be all "left-part" (before delimiter) or "right-part" as requested.
-def split_at(the_str, the_delim, favor_right=True):
-    split_pos = the_str.find(the_delim)
-    if split_pos > 0:
-        left_part  = the_str[0:split_pos]
-        right_part = the_str[split_pos+1:]
-    else:
-        if favor_right:
-            left_part  = None
-            right_part = the_str
-        else:
-            left_part  = the_str
-            right_part = None
-
-    return (left_part, right_part)
-
-# Parse an image reference, return dict containing image reference information
-def parse_image_ref(image_ref):
-    # Image ref:  [registry-and-ns/]repository-name[:tag][@digest]
-    parsed_ref = dict()
-
-    remaining_ref = image_ref
-    at_pos = remaining_ref.rfind("@")
-    if at_pos > 0:
-        parsed_ref["digest"] = remaining_ref[at_pos+1:]
-        remaining_ref = remaining_ref[0:at_pos]
-    else:
-        parsed_ref["digest"] = None
-    colon_pos = remaining_ref.rfind(":")
-    if colon_pos > 0:
-        parsed_ref["tag"] = remaining_ref[colon_pos+1:]
-        remaining_ref = remaining_ref[0:colon_pos]
-    else:
-        parsed_ref["tag"] = None
-    slash_pos = remaining_ref.rfind("/")
-    if slash_pos > 0:
-        parsed_ref["repository"] = remaining_ref[slash_pos+1:]
-        rgy_and_ns = remaining_ref[0:slash_pos]
-    else:
-        parsed_ref["repository"] = remaining_ref
-        rgy_and_ns = "localhost"
-    parsed_ref["registry_and_namespace"] = rgy_and_ns
-
-    rgy, ns = split_at(rgy_and_ns, "/", favor_right=False)
-    if not ns:
-        ns = ""
-
-    parsed_ref["registry"] = rgy
-    parsed_ref["namespace"] = ns
-
-    slash_pos = image_ref.rfind("/")
-    if slash_pos > 0:
-        repo_and_suffix = image_ref[slash_pos+1:]
-    else:
-        repo_and_suffix = image_ref
-    parsed_ref["repository_and_suffix"]  = repo_and_suffix
-
-    return parsed_ref
 
 # Copy chart-templates to a new helmchart directory
 def templateHelmChart(outputDir, helmChart, preservedFiles=None, overwrite=False):
@@ -328,359 +269,6 @@ def addResources(helmChart, csvPath):
     if handleAllFiles:
         logging.error("Found a resource in either the manifest or csv we aren't handling")
         sys.exit(1)
-# Given a resource Kind, return all filepaths of that resource type in a chart directory
-def findTemplatesOfType(helmChart, kind):
-    resources = []
-    for filename in os.listdir(os.path.join(helmChart, "templates")):
-        if filename.endswith(".yaml") or filename.endswith(".yml"):
-            filePath = os.path.join(helmChart, "templates", filename)
-            with open(filePath, 'r') as f:
-                fileYml = yaml.safe_load(f)
-            if fileYml['kind'] == kind:
-                resources.append(filePath)
-            continue
-        else:
-            continue
-    return resources
-
-# For each deployment, identify the image references if any exist in the environment variable fields, insert helm flow control code to reference it, and add image-key to the values.yaml file.
-# If the image-key referenced in the deployment does not exist in `imageMappings` in the Config.yaml, this will fail. Images must be explicitly defined
-def fixEnvVarImageReferences(helmChart, imageKeyMapping):
-    logging.info("Fixing image references in container 'env' section in deployments and values.yaml ...")
-    valuesYaml = os.path.join(helmChart, "values.yaml")
-    with open(valuesYaml, 'r') as f:
-        values = yaml.safe_load(f)
-    deployments = findTemplatesOfType(helmChart, 'Deployment')
-
-    imageKeys = []
-    for deployment in deployments:
-        with open(deployment, 'r') as f:
-            deploy = yaml.safe_load(f)
-        
-        containers = deploy['spec']['template']['spec']['containers']
-        for container in containers:
-            if 'env' not in container: 
-                continue
-            
-            for env in container['env']:
-                image_key = env['name']
-                if image_key.endswith('_IMAGE') == False:
-                    continue
-                image_key = parse_image_ref(env['value'])['repository']
-                try:
-                    image_key = imageKeyMapping[image_key]
-                except KeyError:
-                    logging.critical("No image key mapping provided for imageKey: %s" % image_key)
-                    exit(1)
-                imageKeys.append(image_key)
-                env['value'] = "{{ .Values.global.imageOverrides." + image_key + " }}"
-        with open(deployment, 'w') as f:
-            yaml.dump(deploy, f)
-
-    for imageKey in imageKeys:
-        values['global']['imageOverrides'][imageKey] = ""
-    with open(valuesYaml, 'w') as f:
-        yaml.dump(values, f)
-    logging.info("Image container env references in deployments and values.yaml updated successfully.\n")
-
-# For each deployment, identify the image references if any exist in the image field, insert helm flow control code to reference it, and add image-key to the values.yaml file.
-# If the image-key referenced in the deployment does not exist in `imageMappings` in the Config.yaml, this will fail. Images must be explicitly defined
-def fixImageReferences(helmChart, imageKeyMapping):
-    logging.info("Fixing image and pull policy references in deployments and values.yaml ...")
-    valuesYaml = os.path.join(helmChart, "values.yaml")
-    with open(valuesYaml, 'r') as f:
-        values = yaml.safe_load(f)
-    
-    deployments = findTemplatesOfType(helmChart, 'Deployment')
-    imageKeys = []
-    temp = "" ## temporarily read image ref
-    for deployment in deployments:
-        with open(deployment, 'r') as f:
-            deploy = yaml.safe_load(f)
-        
-        containers = deploy['spec']['template']['spec']['containers']
-        for container in containers:
-            image_key = parse_image_ref(container['image'])["repository"]
-            try:
-                image_key = imageKeyMapping[image_key]
-            except KeyError:
-                logging.critical("No image key mapping provided for imageKey: %s" % image_key)
-                exit(1)
-            imageKeys.append(image_key)
-            # temp = container['image'] 
-            container['image'] = "{{ .Values.global.imageOverrides." + image_key + " }}"
-            container['imagePullPolicy'] = "{{ .Values.global.pullPolicy }}"
-        with open(deployment, 'w') as f:
-            yaml.dump(deploy, f)
-
-    # Remove the placeholder/dummy image overrides we might get from our values template
-    try:
-        del values['global']['imageOverrides']['imageOverride']
-    except KeyError:
-        pass
-    for imageKey in imageKeys:
-        values['global']['imageOverrides'][imageKey] = "" # set to temp to debug
-    with open(valuesYaml, 'w') as f:
-        yaml.dump(values, f)
-    logging.info("Image references and pull policy in deployments and values.yaml updated successfully.\n")
-
-# insers Heml flow control if/end block around a first and last line without changing
-# the indexes of the lines list (so as to not mess up iteration across the lines).
-def insertFlowControlIfAround(lines_list, first_line_index, last_line_index, if_condition):
-   lines_list[first_line_index] = "{{- if %s }}\n%s" % (if_condition, lines_list[first_line_index])
-   lines_list[last_line_index] = "%s{{- end }}\n" % lines_list[last_line_index]
-
-def is_version_compatible(branch, min_release_version, min_backplane_version, min_ocm_version):
-    # Extract the version part from the branch name (e.g., '2.12-integration' -> '2.12')
-    pattern = r'(\d+\.\d+)'  # Matches versions like '2.12'
-    
-    if branch == "main" or branch == "master":
-        return True
-    
-    match = re.search(pattern, branch)
-    if match:
-        v = match.group(1)  # Extract the version
-        branch_version = version.Version(v)  # Create a Version object
-        
-        if "release-ocm" in branch:
-            min_branch_version = version.Version(min_ocm_version)  # Use the minimum release version
-        
-        elif "release" in branch:
-            min_branch_version = version.Version(min_release_version)  # Use the minimum release version
-
-        elif "backplane" in branch or "mce" in branch:
-            min_branch_version = version.Version(min_backplane_version)  # Use the minimum backplane version
-
-        else:
-            logging.error(f"Unrecognized branch type for branch: {branch}")
-            return False
-
-        # Check if the branch version is compatible with the specified minimum branch
-        return branch_version >= min_branch_version
-
-    else:
-        logging.error(f"Version not found in branch: {branch}")
-        return False
-
-# injectHelmFlowControl injects advanced helm flow control which would typically make a .yaml file more difficult to parse. This should be called last.
-def injectHelmFlowControl(deployment, sizes, branch):
-    logging.info("Adding Helm flow control for NodeSelector, Proxy Overrides and SecCompProfile...")
-    deploy = open(deployment, "r")
-    with open(deployment, 'r') as f:
-        deployx = yaml.safe_load(f)
-    lines = deploy.readlines()
-    for i, line in enumerate(lines):
-        if line.strip() == "nodeSelector: \'\'":
-            lines[i] = """{{- with .Values.hubconfig.nodeSelector }}
-      nodeSelector:
-{{ toYaml . | indent 8 }}
-{{- end }}
-"""     
-        if line.strip() == "imagePullSecrets: \'\'":
-            lines[i] = """{{- if .Values.global.pullSecret }}
-      imagePullSecrets:
-      - name: {{ .Values.global.pullSecret }}
-{{- end }}
-"""
-        if line.strip() == "tolerations: \'\'":
-            lines[i] = """{{- with .Values.hubconfig.tolerations }}
-      tolerations:
-      {{- range . }}
-      - {{ if .Key }} key: {{ .Key }} {{- end }}
-        {{ if .Operator }} operator: {{ .Operator }} {{- end }}
-        {{ if .Value }} value: {{ .Value }} {{- end }}
-        {{ if .Effect }} effect: {{ .Effect }} {{- end }}
-        {{ if .TolerationSeconds }} tolerationSeconds: {{ .TolerationSeconds }} {{- end }}
-        {{- end }}
-{{- end }}
-"""
-
-        if line.strip() == "env:" or line.strip() == "env: {}":
-            lines[i] = """        env:
-{{- if .Values.hubconfig.proxyConfigs }}
-        - name: HTTP_PROXY
-          value: {{ .Values.hubconfig.proxyConfigs.HTTP_PROXY }}
-        - name: HTTPS_PROXY
-          value: {{ .Values.hubconfig.proxyConfigs.HTTPS_PROXY }}
-        - name: NO_PROXY
-          value: {{ .Values.hubconfig.proxyConfigs.NO_PROXY }}
-{{- end }}
-"""     
-
-        if is_version_compatible(branch, '9.9', '9.9', '9.9'):
-            if 'replicas:' in line.strip():
-                lines[i] = """  replicas: {{ .Values.hubconfig.replicaCount }}
-"""
-            
-        if sizes:
-            for sizDeployment in sizes["deployments"]:
-                if sizDeployment["name"] == deployx["metadata"]["name"]:
-                    for container in sizDeployment["containers"]:
-                        if line.strip() == "resources: REPLACE-" + container["name"]:
-                            lines[i] = """        resources:
-{{-  if eq .values.hubconfig.hubSize "Small" }}
-          limits:
-            cpu: """ + container["Small"]["limits"]["cpu"] + """
-            memory: """ + container["Small"]["limits"]["memory"] + """
-          requests:
-            cpu: """ + container["Small"]["requests"]["cpu"] + """
-            memory: """ + container["Small"]["requests"]["memory"] + """
-{{- end }}
-{{ if eq .values.hubconfig.hubSize "Medium" }}
-          limits:
-            cpu: """ + container["Medium"]["limits"]["cpu"] + """
-            memory: """ + container["Medium"]["limits"]["memory"] + """
-          requests:
-            cpu: """ + container["Medium"]["requests"]["cpu"] + """
-            memory: """ + container["Medium"]["requests"]["memory"] + """
-{{- end }}
-{{-  if eq .values.hubconfig.hubSize "Large" }}
-          limits:
-            cpu: """ + container["Large"]["limits"]["cpu"] + """
-            memory: """ + container["Large"]["limits"]["memory"] + """
-          requests:
-            cpu: """ + container["Large"]["requests"]["cpu"] + """
-            memory: """ + container["Large"]["requests"]["memory"] + """
-{{- end }}
-{{ if eq .values.hubconfig.hubSize "ExtraLarge" }}
-          limits:
-            cpu: """ + container["ExtraLarge"]["limits"]["cpu"] + """
-            memory: """ + container["ExtraLarge"]["limits"]["memory"] + """
-          requests:
-            cpu: """ + container["ExtraLarge"]["requests"]["cpu"] + """
-            memory: """ + container["ExtraLarge"]["requests"]["memory"] + """
-{{- end }}
-"""
-        if line.strip() == "seccompProfile:":
-            next_line = lines[i+1]  # Ignore possible reach beyond end-of-list, not really possible
-            if next_line.strip() == "type: RuntimeDefault":
-                insertFlowControlIfAround(lines, i, i+1, "semverCompare \">=4.11.0\" .Values.hubconfig.ocpVersion")
-                if is_version_compatible(branch, '9.9', '2.7', '2.12'):
-                    insertFlowControlIfAround(lines, i, i+1, ".Values.global.deployOnOCP")
-    #
-    a_file = open(deployment, "w")
-    a_file.writelines(lines)
-    a_file.close()
-    logging.info("Added Helm flow control for NodeSelector, Proxy Overrides and SecCompProfile.\n")
-
-# updateDeployments adds standard configuration to the deployments (antiaffinity, security policies, and tolerations)
-def updateDeployments(helmChart, operator, exclusions, sizes, branch):
-    logging.info("Updating deployments with antiaffinity, security policies, and tolerations ...")
-    deploySpecYaml = os.path.join(os.path.dirname(os.path.realpath(__file__)), "chart-templates/templates/deploymentspec.yaml")
-    with open(deploySpecYaml, 'r') as f:
-        deploySpec = yaml.safe_load(f)
-    deployments = findTemplatesOfType(helmChart, 'Deployment')
-    for deployment in deployments:
-        with open(deployment, 'r') as f:
-            deploy = yaml.safe_load(f)
-        affinityList = deploySpec['affinity']['podAntiAffinity']['preferredDuringSchedulingIgnoredDuringExecution']
-        for antiaffinity in affinityList:
-            antiaffinity['podAffinityTerm']['labelSelector']['matchExpressions'][0]['values'][0] = deploy['metadata']['name']
-
-        pod_template = deploy['spec']['template']
-        pod_template['metadata']['labels']['ocm-antiaffinity-selector'] = deploy['metadata']['name']
-        if sizes:
-            for  sizDeployment in sizes["deployments"]:
-                if sizDeployment["name"] == deploy["metadata"]["name"]:
-                    for i in deploy['spec']['template']['spec']['containers']:            
-                        if not any(d['name'] == i['name'] for d in sizDeployment["containers"]):
-                            logging.error("Missing container in sizes.yaml")
-                            exit(1)
-                        for sizContainer in sizDeployment["containers"]:
-                            if sizContainer["name"] == i["name"]:
-                                i['resources'] = 'REPLACE-' + i['name']
-
-        pod_template_spec = pod_template['spec']
-        pod_template_spec['affinity'] = deploySpec['affinity']
-        pod_template_spec['tolerations'] = ''
-        pod_template_spec['hostNetwork'] = False
-        pod_template_spec['hostPID'] = False
-        pod_template_spec['hostIPC'] = False
-        
-        # Set automountServiceAccountToken only if is configured for the operator.
-        if 'automountServiceAccountToken' in operator:
-            automountSAToken = operator.get('automountServiceAccountToken')
-            if isinstance(automountSAToken, bool):
-                pod_template_spec['automountServiceAccountToken'] = operator.get('automountServiceAccountToken')
-            else:
-                logging.warning("automountServiceAccountToken should be a boolean. Ignoring invalid value.")
-
-        if 'securityContext' not in pod_template_spec:
-            pod_template_spec['securityContext'] = {}
-        pod_security_context = pod_template_spec['securityContext']
-        pod_security_context['runAsNonRoot'] = True
-
-        if 'seccompProfile' not in pod_security_context:
-            pod_security_context['seccompProfile'] = {'type': 'RuntimeDefault'}
-            # This will be made conditional on OCP version >= 4.11 by injectHelmFlowControl()
-        else:
-            if pod_security_context['seccompProfile']['type'] != 'RuntimeDefault':
-                logging.warning("Leaving non-standard pod-level seccompprofile setting.")
-
-        pod_template_spec['nodeSelector'] = ""
-        pod_template_spec['imagePullSecrets'] = ''
-
-        containers = pod_template_spec['containers']
-        for container in containers:
-            if 'env' not in container:
-                container['env'] = {}
-
-            if 'securityContext' not in container: 
-                container['securityContext'] = {}
-            container_security_context = container['securityContext']
-            container_security_context['allowPrivilegeEscalation'] = False
-            container_security_context['capabilities'] = {'drop': ['ALL']}
-            container_security_context['privileged'] = False
-            if 'readOnlyRootFilesystem' not in exclusions:
-                container_security_context['readOnlyRootFilesystem'] = True
-
-            if 'seccompProfile' in container_security_context:
-                if container_security_context['seccompProfile']['type'] == 'RuntimeDefault':
-                    # Remove, to allow pod-level setting to have effect.
-                    del container_security_context['seccompProfile']
-                else:
-                    container_name = container['name']
-                    logging.warning("Leaving non-standard seccompprofile setting for container %s" % container_name)
-
-        
-        with open(deployment, 'w') as f:
-            yaml.dump(deploy, f)
-        logging.info("Deployments updated with antiaffinity, security policies, and tolerations successfully. \n")
-
-        injectHelmFlowControl(deployment, sizes, branch)
-
-# updateRBAC adds standard configuration to the RBAC resources (clusterroles, roles, clusterrolebindings, and rolebindings)
-def updateRBAC(helmChart):
-    logging.info("Updating clusterroles, roles, clusterrolebindings, and rolebindings ...")
-    clusterroles = findTemplatesOfType(helmChart, 'ClusterRole')
-    roles = findTemplatesOfType(helmChart, 'Role')
-    clusterrolebindings = findTemplatesOfType(helmChart, 'ClusterRoleBinding')
-    rolebindings = findTemplatesOfType(helmChart, 'RoleBinding')
-
-    for rbacFile in clusterroles + roles + clusterrolebindings + rolebindings:
-        with open(rbacFile, 'r') as f:
-            rbac = yaml.safe_load(f)
-        rbac['metadata']['name'] = "{{ .Values.org }}:{{ .Chart.Name }}:" + rbac['metadata']['name']
-        if rbac['kind'] in ['RoleBinding', 'ClusterRoleBinding']:
-            rbac['roleRef']['name'] = "{{ .Values.org }}:{{ .Chart.Name }}:" + rbac['roleRef']['name']
-        with open(rbacFile, 'w') as f:
-            yaml.dump(rbac, f)
-    logging.info("Clusterroles, roles, clusterrolebindings, and rolebindings updated. \n")
-
-
-def injectRequirements(helmChart, operator, exclusions, sizes, branch):
-    logging.info("Updating Helm chart '%s' with onboarding requirements ...", helmChart)
-    imageKeyMapping = operator.get('imageMappings', {})
-
-    # Fixes image references in the Helm chart.
-    fixImageReferences(helmChart, imageKeyMapping)
-    fixEnvVarImageReferences(helmChart, imageKeyMapping)
-
-    # Updates RBAC and deployment configuration in the Helm chart.
-    updateRBAC(helmChart)
-    updateDeployments(helmChart, operator, exclusions, sizes, branch)
-
-    logging.info("Updated Chart '%s' successfully\n", helmChart)
 
 def addCMAs(repo, operator, outputDir):
     if 'bundlePath' in operator:
@@ -898,6 +486,7 @@ def main():
             repository = Repo.clone_from(repo["github_ref"], repo_path) # Clone repo to above path
             if 'branch' in repo:
                 repository.git.checkout(repo['branch']) # If a branch is specified, checkout that branch
+
             sizesyaml = repo_path + "/bundle/manifests/sizes.yaml"
             if os.path.isfile(sizesyaml):
                 with open(sizesyaml, 'r') as f:
@@ -1026,7 +615,7 @@ def main():
             if not skipOverrides:
                 logging.info("Adding Overrides to helm chart '%s' (set --skipOverrides=true to skip) ...", operator["name"])
                 exclusions = operator["exclusions"] if "exclusions" in operator else []
-                injectRequirements(helmChart, operator, exclusions, sizes, branch)
+                common.inject_requirements(helmChart, operator, exclusions, sizes, branch)
                 logging.info("Overrides added to helm chart '%s' successfully.", operator["name"])
 
     logging.info("All repositories and operators processed successfully.")
