+++ bundle-generation/generate-charts.py
@@ -10,7 +10,8 @@ import yaml
 import logging
 import coloredlogs
 import subprocess
-import re
+import common
+
 from git import Repo, exc
 from packaging import version
 
@@ -19,53 +20,8 @@ from validate_csv import *
 # Configure logging with coloredlogs
 coloredlogs.install(level='DEBUG')  # Set the logging level as needed
 
-# Parse an image reference, return dict containing image reference information
-def parse_image_ref(image_ref):
-   # Image ref:  [registry-and-ns/]repository-name[:tag][@digest]
-   parsed_ref = dict()
-
-   remaining_ref = image_ref
-   at_pos = remaining_ref.rfind("@")
-   if at_pos > 0:
-      parsed_ref["digest"] = remaining_ref[at_pos+1:]
-      remaining_ref = remaining_ref[0:at_pos]
-   else:
-      parsed_ref["digest"] = None
-   colon_pos = remaining_ref.rfind(":")
-   if colon_pos > 0:
-      parsed_ref["tag"] = remaining_ref[colon_pos+1:]
-      remaining_ref = remaining_ref[0:colon_pos]
-   else:
-      parsed_ref["tag"] = None
-   slash_pos = remaining_ref.rfind("/")
-   if slash_pos > 0:
-      parsed_ref["repository"] = remaining_ref[slash_pos+1:]
-      rgy_and_ns = remaining_ref[0:slash_pos]
-   else:
-      parsed_ref["repository"] = remaining_ref
-      rgy_and_ns = "localhost"
-   parsed_ref["registry_and_namespace"] = rgy_and_ns
-
-   rgy, ns = split_at(rgy_and_ns, "/", favor_right=False)
-   if not ns:
-      ns = ""
-
-   parsed_ref["registry"] = rgy
-   parsed_ref["namespace"] = ns
-
-   slash_pos = image_ref.rfind("/")
-   if slash_pos > 0:
-      repo_and_suffix = image_ref[slash_pos+1:]
-   else:
-      repo_and_suffix = image_ref
-   parsed_ref["repository_and_suffix"]  = repo_and_suffix
-
-   return parsed_ref
-
-
 def updateAddOnDeploymentConfig(yamlContent):
     yamlContent['metadata']['namespace'] = '{{ .Values.global.namespace }}'
-
 
 def updateClusterManagementAddOn(yamlContent):
     if 'spec' not in yamlContent:
@@ -115,7 +71,7 @@ def updateClusterRoleBinding(yamlContent):
         sub['namespace'] = '{{ .Values.global.namespace }}'
 
 def escapeTemplateVariables(helmChart, variables):
-    addonTemplates = findTemplatesOfType(helmChart, 'AddOnTemplate')
+    addonTemplates = common.find_templates_of_type(helmChart, 'AddOnTemplate')
     for addonTemplate in addonTemplates:
         for variable in variables:
             logging.info("Start to escape vriable %s", variable)
@@ -280,438 +236,6 @@ def copyHelmChart(destinationChartPath, repo, chart, chartVersion):
     shutil.copyfile(os.path.join(os.path.dirname(os.path.realpath(__file__)), "chart-templates", "values.yaml"), os.path.join(destinationChartPath, "values.yaml"))
 
     logging.info(f"Finished processing chart: '{chartName}'\n")
-
-# Given a resource Kind, return all filepaths of that resource type in a chart directory
-def findTemplatesOfType(helmChart, kind):
-    resources = []
-    for filename in os.listdir(os.path.join(helmChart, "templates")):
-        if filename.endswith(".yaml") or filename.endswith(".yml"):
-            filePath = os.path.join(helmChart, "templates", filename)
-            with open(filePath, 'r') as f:
-                fileYml = yaml.safe_load(f)
-            if fileYml['kind'] == kind:
-                resources.append(filePath)
-            continue
-        else:
-            continue
-    return resources
-
-# For each deployment, identify the image references if any exist in the environment variable fields, insert helm flow control code to reference it, and add image-key to the values.yaml file.
-# If the image-key referenced in the deployment does not exist in `imageMappings` in the Config.yaml, this will fail. Images must be explicitly defined
-def fixEnvVarImageReferences(helmChart, imageKeyMapping):
-    logging.info("Fixing image references in container 'env' section in deployments and values.yaml ...")
-    
-    # Path to the values.yaml file
-    valuesYaml = os.path.join(helmChart, "values.yaml")
-
-    # Check if the values.yaml file exists
-    if not os.path.exists(valuesYaml):
-        logging.error(f"{valuesYaml} does not exist. Skipping environment variable image reference updates.")
-        return
-
-    with open(valuesYaml, 'r') as f:
-        values = yaml.safe_load(f)
-    deployments = findTemplatesOfType(helmChart, 'Deployment')
-
-    imageKeys = []
-    for deployment in deployments:
-        with open(deployment, 'r') as f:
-            deploy = yaml.safe_load(f)
-        
-        containers = deploy['spec']['template']['spec']['containers']
-        for container in containers:
-            if 'env' not in container: 
-                continue
-            
-            for env in container['env']:
-                image_key = env['name']
-                if image_key.endswith('_IMAGE') == False:
-                    continue
-                image_key = parse_image_ref(env['value'])['repository']
-                try:
-                    image_key = imageKeyMapping[image_key]
-                except KeyError:
-                    logging.critical("No image key mapping provided for imageKey: %s" % image_key)
-                    exit(1)
-                imageKeys.append(image_key)
-                env['value'] = "{{ .Values.global.imageOverrides." + image_key + " }}"
-        with open(deployment, 'w') as f:
-            yaml.dump(deploy, f, width=float("inf"))
-
-    for imageKey in imageKeys:
-        values['global']['imageOverrides'][imageKey] = ""
-    with open(valuesYaml, 'w') as f:
-        yaml.dump(values, f, width=float("inf"))
-    logging.info("Image container env references in deployments and values.yaml updated successfully.\n")
-
-# For each deployment, identify the image references if any exist in the image field, insert helm flow control code to reference it, and add image-key to the values.yaml file.
-# If the image-key referenced in the deployment does not exist in `imageMappings` in the Config.yaml, this will fail. Images must be explicitly defined
-def fixImageReferences(helmChart, imageKeyMapping):
-    logging.info("Fixing image and pull policy references in deployments and values.yaml ...")
-
-    # Path to the values.yaml file
-    valuesYaml = os.path.join(helmChart, "values.yaml")
-
-    # Check if the values.yaml file exists
-    if not os.path.exists(valuesYaml):
-        logging.error(f"{valuesYaml} does not exist. Skipping image and pull policy updates.")
-        return  # Exit the function if the file doesn't exist
-
-    with open(valuesYaml, 'r') as f:
-        values = yaml.safe_load(f)
-    
-    deployments = findTemplatesOfType(helmChart, 'Deployment')
-    imageKeys = []
-    temp = "" ## temporarily read image ref
-    for deployment in deployments:
-        with open(deployment, 'r') as f:
-            deploy = yaml.safe_load(f)
-        
-        containers = deploy['spec']['template']['spec']['containers']
-        for container in containers:
-            image_key = parse_image_ref(container['image'])["repository"]
-            try:
-                image_key = imageKeyMapping[image_key]
-            except KeyError:
-                logging.critical("No image key mapping provided for imageKey: %s" % image_key)
-                exit(1)
-            imageKeys.append(image_key)
-            # temp = container['image'] 
-            container['image'] = "{{ .Values.global.imageOverrides." + image_key + " }}"
-            container['imagePullPolicy'] = "{{ .Values.global.pullPolicy }}"
-            args = container['args']
-            refreshed_args = []
-            for arg in args:
-                if "--agent-image-name" not in arg:
-                    refreshed_args.append(arg)
-                else:
-                    refreshed_args.append("--agent-image-name="+"{{ .Values.global.imageOverrides." + image_key + " }}")
-            container['args'] = refreshed_args
-        with open(deployment, 'w') as f:
-            yaml.dump(deploy, f, width=float("inf"))
-
-    if 'imageOverride' in values['global']['imageOverrides']:
-        del values['global']['imageOverrides']['imageOverride']
-
-    for imageKey in imageKeys:
-        values['global']['imageOverrides'][imageKey] = "" # set to temp to debug
-
-    with open(valuesYaml, 'w') as f:
-        yaml.dump(values, f, width=float("inf"))
-    logging.info("Image references and pull policy in deployments and values.yaml updated successfully.\n")
-
-# insers Heml flow control if/end block around a first and last line without changing
-# the indexes of the lines list (so as to not mess up iteration across the lines).
-def insertFlowControlIfAround(lines_list, first_line_index, last_line_index, if_condition):
-   lines_list[first_line_index] = "{{- if %s }}\n%s" % (if_condition, lines_list[first_line_index])
-   lines_list[last_line_index] = "%s{{- end }}\n" % lines_list[last_line_index]
-
-def is_version_compatible(branch, min_release_version, min_backplane_version, min_ocm_version):
-    # Extract the version part from the branch name (e.g., '2.12-integration' -> '2.12')
-    pattern = r'(\d+\.\d+)'  # Matches versions like '2.12'
-    
-    if branch == "main" or branch == "master":
-        return True
-    
-    match = re.search(pattern, branch)
-    if match:
-        v = match.group(1)  # Extract the version
-        branch_version = version.Version(v)  # Create a Version object
-        
-        if "release-ocm" in branch:
-            min_branch_version = version.Version(min_ocm_version)  # Use the minimum release version
-        
-        elif "release" in branch:
-            min_branch_version = version.Version(min_release_version)  # Use the minimum release version
-
-        elif "backplane" in branch or "mce" in branch:
-            min_branch_version = version.Version(min_backplane_version)  # Use the minimum backplane version
-
-        else:
-            logging.error(f"Unrecognized branch type for branch: {branch}")
-            return False
-
-        # Check if the branch version is compatible with the specified minimum branch
-        return branch_version >= min_branch_version
-
-    else:
-        return False
-
-# injectHelmFlowControl injects advanced helm flow control which would typically make a .yaml file more difficult to parse. This should be called last.
-def injectHelmFlowControl(deployment, branch):
-    logging.info("Adding Helm flow control for NodeSelector, Proxy Overrides, and SeccompProfile ...")
-    deploy = open(deployment, "r")
-    lines = deploy.readlines()
-    for i, line in enumerate(lines):
-        if line.strip() == "nodeSelector: \'\'":
-            lines[i] = """{{- with .Values.hubconfig.nodeSelector }}
-      nodeSelector:
-{{ toYaml . | indent 8 }}
-{{- end }}
-"""     
-        if line.strip() == "imagePullSecrets: \'\'":
-            lines[i] = """{{- if .Values.global.pullSecret }}
-      imagePullSecrets:
-      - name: {{ .Values.global.pullSecret }}
-{{- end }}
-"""
-        if line.strip() == "tolerations: \'\'":
-            lines[i] = """{{- with .Values.hubconfig.tolerations }}
-      tolerations:
-      {{- range . }}
-      - {{ if .Key }} key: {{ .Key }} {{- end }}
-        {{ if .Operator }} operator: {{ .Operator }} {{- end }}
-        {{ if .Value }} value: {{ .Value }} {{- end }}
-        {{ if .Effect }} effect: {{ .Effect }} {{- end }}
-        {{ if .TolerationSeconds }} tolerationSeconds: {{ .TolerationSeconds }} {{- end }}
-        {{- end }}
-{{- end }}
-"""
-
-
-        if line.strip() == "env:" or line.strip() == "env: {}":
-            lines[i] = """        env:
-{{- if .Values.hubconfig.proxyConfigs }}
-        - name: HTTP_PROXY
-          value: {{ .Values.hubconfig.proxyConfigs.HTTP_PROXY }}
-        - name: HTTPS_PROXY
-          value: {{ .Values.hubconfig.proxyConfigs.HTTPS_PROXY }}
-        - name: NO_PROXY
-          value: {{ .Values.hubconfig.proxyConfigs.NO_PROXY }}
-{{- end }}
-"""
-
-        if is_version_compatible(branch, '9.9', '9.9', '9.9'):
-            if 'replicas:' in line.strip():
-                lines[i] = """  replicas: {{ .Values.hubconfig.replicaCount }}
-"""
-
-        if line.strip() == "seccompProfile:":
-            next_line = lines[i+1]  # Ignore possible reach beyond end-of-list, not really possible
-            prev_line = lines[i-1]
-            if next_line.strip() == "type: RuntimeDefault" and "semverCompare" not in prev_line:
-                insertFlowControlIfAround(lines, i, i+1, "semverCompare \">=4.11.0\" .Values.hubconfig.ocpVersion")
-                if is_version_compatible(branch, '9.9', '2.7', '2.12'):
-                    insertFlowControlIfAround(lines, i, i+1, ".Values.global.deployOnOCP")
-
-        a_file = open(deployment, "w")
-        a_file.writelines(lines)
-        a_file.close()
-    logging.info("Added Helm flow control for NodeSelector, Proxy, and SeccompProfile Overrides.\n")
-
-def addPullSecretOverride(deployment):
-    deploy = open(deployment, "r")
-    lines = deploy.readlines()
-    for i, line in enumerate(lines):
-        if line.strip() == "env:" or line.strip() == "env: {}":
-            logging.info("Adding image pull secret environment variable to managed-serviceaccount deployment")
-            lines[i] = """        env:
-{{- if .Values.global.pullSecret }}
-        - name: AGENT_IMAGE_PULL_SECRET
-          value: {{ .Values.global.pullSecret }}
-{{- end }}
-"""
-        a_file = open(deployment, "w")
-        a_file.writelines(lines)
-        a_file.close()
-
-# updateDeployments adds standard configuration to the deployments (antiaffinity, security policies, and tolerations)
-def updateDeployments(chartName, helmChart, exclusions, inclusions, branch):
-    logging.info("Updating deployments with antiaffinity, security policies, and tolerations ...")
-    deploySpecYaml = os.path.join(os.path.dirname(os.path.realpath(__file__)), "chart-templates/templates/deploymentspec.yaml")
-    with open(deploySpecYaml, 'r') as f:
-        deploySpec = yaml.safe_load(f)
-    
-    deployments = findTemplatesOfType(helmChart, 'Deployment')
-    for deployment in deployments:
-        with open(deployment, 'r') as f:
-            deploy = yaml.safe_load(f)
-        deploy['metadata'].pop('namespace')
-        affinityList = deploySpec['affinity']['podAntiAffinity']['preferredDuringSchedulingIgnoredDuringExecution']
-        for antiaffinity in affinityList:
-            antiaffinity['podAffinityTerm']['labelSelector']['matchExpressions'][0]['values'][0] = deploy['metadata']['name']
-        deploy['spec']['template']['spec']['affinity'] = deploySpec['affinity']
-        deploy['spec']['template']['spec']['tolerations'] = ''
-        deploy['spec']['template']['spec']['hostNetwork'] = False
-        deploy['spec']['template']['spec']['hostPID'] = False
-        deploy['spec']['template']['spec']['hostIPC'] = False
-        if 'securityContext' not in deploy['spec']['template']['spec']:
-            deploy['spec']['template']['spec']['securityContext'] = {}
-        deploy['spec']['template']['spec']['securityContext']['runAsNonRoot'] = True
-        deploy['spec']['template']['metadata']['labels']['ocm-antiaffinity-selector'] = deploy['metadata']['name']
-        deploy['spec']['template']['spec']['nodeSelector'] = ""
-        deploy['spec']['template']['spec']['imagePullSecrets'] = ''
-        pod_template_spec = deploy['spec']['template']['spec']
-        if 'securityContext' not in pod_template_spec:
-            pod_template_spec['securityContext'] = {}
-        pod_security_context = pod_template_spec['securityContext']
-        pod_security_context['runAsNonRoot'] = True
-        if 'seccompProfile' not in pod_security_context:
-            pod_security_context['seccompProfile'] = {'type': 'RuntimeDefault'}
-            # This will be made conditional on OCP version >= 4.11 by injectHelmFlowControl()
-        else:
-            if pod_security_context['seccompProfile']['type'] != 'RuntimeDefault':
-                logging.warning("Leaving non-standard pod-level seccompprofile setting.")
-
-        containers = deploy['spec']['template']['spec']['containers']
-        for container in containers:
-            if 'securityContext' not in container: 
-                container['securityContext'] = {}
-            if 'env' not in container: 
-                container['env'] = {}
-            container['securityContext']['allowPrivilegeEscalation'] = False
-            container['securityContext']['capabilities'] = {}
-            container['securityContext']['capabilities']['drop'] = ['ALL']
-            container['securityContext']['privileged'] = False
-            container['securityContext']['runAsNonRoot'] = True
-            if 'readOnlyRootFilesystem' not in exclusions:
-                container['securityContext']['readOnlyRootFilesystem'] = True
-            if 'seccompProfile' in container['securityContext']:
-                if container['securityContext']['seccompProfile']['type'] == 'RuntimeDefault':
-                    # Remove, to allow pod-level setting to have effect.
-                    del container['securityContext']['seccompProfile']
-                else:
-                    container_name = container['name']
-                    logging.warning("Leaving non-standard seccompprofile setting for container %s" % container_name)
-        
-        with open(deployment, 'w') as f:
-            yaml.dump(deploy, f, width=float("inf"))
-        logging.info("Deployments updated with antiaffinity, security policies, and tolerations successfully. \n")
-
-        injectHelmFlowControl(deployment, branch)
-        if 'pullSecretOverride' in inclusions:
-            addPullSecretOverride(deployment)
-
-# injectAnnotationsForAddonTemplate injects following annotations for deployments in the AddonTemplate:
-# - target.workload.openshift.io/management: '{"effect": "PreferredDuringScheduling"}'
-def injectAnnotationsForAddonTemplate(helmChart):
-    logging.info("Injecting Annotations for deployments in the AddonTemplate ...")
-
-    addonTemplates = findTemplatesOfType(helmChart, 'AddOnTemplate')
-    for addonTemplate in addonTemplates:
-        injected = False
-        with open(addonTemplate, 'r') as f:
-            templateContent = yaml.safe_load(f)
-            agentSpec = templateContent['spec']['agentSpec']
-            if 'workload' not in agentSpec:
-                return
-            workload = agentSpec['workload']
-            if 'manifests' not in workload:
-                return
-            manifests = workload['manifests']
-            for manifest in manifests:
-                if manifest['kind'] == 'Deployment':
-                    metadata = manifest['spec']['template']['metadata']
-                    if 'annotations' not in metadata:
-                        metadata['annotations'] = {}
-                    if 'target.workload.openshift.io/management' not in metadata['annotations']:
-                        metadata['annotations']['target.workload.openshift.io/management'] = '{"effect": "PreferredDuringScheduling"}'
-                        injected = True
-        if injected:
-            with open(addonTemplate, 'w') as f:
-                yaml.dump(templateContent, f, width=float("inf"))
-                logging.info("Annotations injected successfully. \n")
-
-
-# fixImageReferencesForAddonTemplate identify the image references for every deployment in addontemplates, if any exist
-# in the image field, insert helm flow control code to reference it, and add image-key to the values.yaml file.
-# If the image-key referenced in the addon template deployment does not exist in `imageMappings` in the Config.yaml,
-# this will fail. Images must be explicitly defined
-def fixImageReferencesForAddonTemplate(helmChart, imageKeyMapping):
-    logging.info("Fixing image references in addon templates and values.yaml ...")
-
-    addonTemplates = findTemplatesOfType(helmChart, 'AddOnTemplate')
-    imageKeys = []
-    temp = "" ## temporarily read image ref
-    for addonTemplate in addonTemplates:
-        with open(addonTemplate, 'r') as f:
-            templateContent = yaml.safe_load(f)
-            agentSpec = templateContent['spec']['agentSpec']
-            if 'workload' not in agentSpec:
-                return
-            workload = agentSpec['workload']
-            if 'manifests' not in workload:
-                return
-            manifests = workload['manifests']
-            imageKeys = []
-            for manifest in manifests:
-                if manifest['kind'] == 'Deployment':
-                    containers = manifest['spec']['template']['spec']['containers']
-                    for container in containers:
-                        image_key = parse_image_ref(container['image'])["repository"]
-                        try:
-                            image_key = imageKeyMapping[image_key]
-                        except KeyError:
-                            logging.critical("No image key mapping provided for imageKey: %s" % image_key)
-                            exit(1)
-                        imageKeys.append(image_key)
-                        container['image'] = "{{ .Values.global.imageOverrides." + image_key + " }}"
-                        # container['imagePullPolicy'] = "{{ .Values.global.pullPolicy }}"
-        with open(addonTemplate, 'w') as f:
-            yaml.dump(templateContent, f, width=float("inf"))
-            logging.info("AddOnTemplate updated with image override successfully. \n")
-
-    if len(imageKeys) == 0:
-        return
-    valuesYaml = os.path.join(helmChart, "values.yaml")
-    with open(valuesYaml, 'r') as f:
-        values = yaml.safe_load(f)
-    if 'imageOverride' in values['global']['imageOverrides']:
-        del values['global']['imageOverrides']['imageOverride']
-    for imageKey in imageKeys:
-        values['global']['imageOverrides'][imageKey] = "" # set to temp to debug
-    with open(valuesYaml, 'w') as f:
-        yaml.dump(values, f, width=float("inf"))
-    logging.info("Image references and pull policy in addon templates and values.yaml updated successfully.\n")
-
-
-# updateRBAC adds standard configuration to the RBAC resources (clusterroles, roles, clusterrolebindings, and rolebindings)
-def updateRBAC(helmChart, chartName):
-    logging.info("Updating clusterroles, roles, clusterrolebindings, and rolebindings ...")
-    clusterroles = findTemplatesOfType(helmChart, 'ClusterRole')
-    roles = findTemplatesOfType(helmChart, 'Role')
-    clusterrolebindings = findTemplatesOfType(helmChart, 'ClusterRoleBinding')
-    rolebindings = findTemplatesOfType(helmChart, 'RoleBinding')
-
-    for rbacFile in clusterroles + roles + clusterrolebindings + rolebindings:
-        with open(rbacFile, 'r') as f:
-            rbac = yaml.safe_load(f)
-        rbac['metadata']['name'] = "{{ .Values.org }}:{{ .Chart.Name }}:" + chartName
-        if rbac['kind'] in ['RoleBinding', 'ClusterRoleBinding']:
-            rbac['roleRef']['name'] = "{{ .Values.org }}:{{ .Chart.Name }}:" + chartName
-        with open(rbacFile, 'w') as f:
-            yaml.dump(rbac, f, width=float("inf"))
-    logging.info("Clusterroles, roles, clusterrolebindings, and rolebindings updated. \n")
-
-
-def injectRequirements(helmChart, chartName, imageKeyMapping, skipRBACOverrides, exclusions, inclusions, branch):
-    logging.info("Updating Helm chart '%s' with onboarding requirements ...", helmChart)
-    fixImageReferences(helmChart, imageKeyMapping)
-    fixEnvVarImageReferences(helmChart, imageKeyMapping)
-    fixImageReferencesForAddonTemplate(helmChart, imageKeyMapping)
-    injectAnnotationsForAddonTemplate(helmChart)
-
-    if not skipRBACOverrides:
-        updateRBAC(helmChart, chartName)
-    updateDeployments(chartName, helmChart, exclusions, inclusions, branch)
-
-    logging.info("Updated Chart '%s' successfully", helmChart)
-
-def split_at(the_str, the_delim, favor_right=True):
-
-   split_pos = the_str.find(the_delim)
-   if split_pos > 0:
-      left_part  = the_str[0:split_pos]
-      right_part = the_str[split_pos+1:]
-   else:
-      if favor_right:
-         left_part  = None
-         right_part = the_str
-      else:
-         left_part  = the_str
-         right_part = None
-
-   return (left_part, right_part)
 
 def addCRDs(repo, chart, outputDir):
     if not 'chart-path' in chart:
@@ -899,7 +423,7 @@ def main():
                 inclusions = chart.get("inclusions", [])
                 skip_rbac_overrides = chart.get("skipRBACOverrides", False)
 
-                injectRequirements(destinationChartPath, chart_name, image_mappings, skip_rbac_overrides, exclusions,
+                common.inject_requirements(destinationChartPath, chart_name, chart, image_mappings, skip_rbac_overrides, exclusions,
                                    inclusions, branch)
                 logging.info("Overrides added.\n")
